feat: eventbus: log error on slow consumers (#3031)
* feat: eventbus: log error on slow consumers

Slow consumers can stall libp2p in hard to debug ways. For example, a
slow consumer of Identify events can prevent NewStreams from opening. In
some cases we do want this back-pressure, but in other cases it is a bug
in the user's application (or even in go-libp2p). The recommended
approach has been to use metrics with Grafana and Prometheus to identify
these issues. In practice, that's been a non-trivial task for users to
setup. A simple log would help identify these issues.

Note that the codepath is relatively unchanged in the normal case. Only
if a producer is stalled will the extra work be put in.

Co-authored-by: Wondertan <hlibwondertan@gmail.com>

* fix: eventbus: fix deadlock in closing a stalled wildcard subscriber

* Hold lock when setting slowConsumer timer

---------

Co-authored-by: Wondertan <hlibwondertan@gmail.com>